requirements.txt
sqlmodel
feedparser
loguru
python-dateutil
requests
